+++ src/toad/screens/store.py
@@ -1,8 +1,6 @@
 from contextlib import suppress
 from dataclasses import dataclass
 from itertools import zip_longest
-import json
-import os
 from pathlib import Path
 from random import shuffle
 from typing import Literal, Self
@@ -17,14 +15,17 @@ from textual.app import ComposeResult
 from textual.content import Content
 from textual.css.query import NoMatches
 from textual.message import Message
+from textual.reactive import reactive
 from textual import containers
 from textual import widgets
 
 import toad
 from toad.app import ToadApp
-from toad.db import DB
 from toad.pill import pill
+from toad import messages
+from toad.widgets.directory_input import DirectoryInput
 from toad.widgets.mandelbrot import Mandelbrot
+from toad.widgets.condensed_path import CondensedPath
 from toad.widgets.grid_select import GridSelect
 from toad.agent_schema import Agent
 from toad.agents import read_agents
@@ -47,10 +48,82 @@ QR = """\
 
 
 @dataclass
-class LaunchAgent(Message):
-    identity: str
-    session_id: str | None = None
-    pk: int | None = None
+class ChangeDirectory(Message):
+    path: str
+
+
+class DirectoryDisplay(containers.HorizontalGroup):
+
+    BINDINGS = [("escape", "dismiss", "Dismiss")]
+
+    DEFAULT_CSS = """
+    DirectoryDisplay {
+        CondensedPath { display: block; }
+        DirectoryInput { display: none; }
+        &.-edit {
+            CondensedPath { display: none}
+            DirectoryInput { display: block; }
+        }
+    }
+    """
+
+    project_dir: reactive[Path] = reactive(Path)
+    path = reactive("")
+    edit = reactive(False, toggle_class="-edit")
+
+    directory_input = getters.query_one(DirectoryInput)
+
+    def __init__(self, project_dir: Path) -> None:
+        super().__init__()
+        self.project_dir = project_dir
+        self.path = str(project_dir)
+
+    def watch_project_dir(self, path: Path) -> None:
+        self.path = str(path)
+        self.refresh(layout=True)
+
+    def focus(self, scroll_visible=True) -> Self:
+        self.edit = True
+        self.directory_input.focus(scroll_visible=scroll_visible)
+        return self
+
+    @on(events.Click, "CondensedPath")
+    def on_click(self) -> None:
+        self.edit = True
+        self.directory_input.focus()
+
+    @on(events.DescendantBlur)
+    def on_blur(self):
+        self.action_dismiss()
+
+    @on(widgets.Input.Submitted)
+    def on_input_submitted(self, event: widgets.Input.Submitted) -> None:
+        self.edit = False
+        path = Path(event.value)
+        if not path.is_dir():
+            self.notify(
+                f"Unable to change directory to {str(path)!r}",
+                title="Change directory",
+                severity="error",
+            )
+            return
+        path = path.expanduser().resolve()
+        self.post_message(ChangeDirectory(str(path)))
+
+    def action_dismiss(self) -> None:
+        self.edit = False
+        self.directory_input.value = self.path
+
+    def watch_edit(self, edit: bool) -> None:
+        if not edit and self.directory_input.has_focus:
+            self.directory_input.blur()
+
+    def compose(self) -> ComposeResult:
+        yield widgets.Label("📁 ")
+        yield CondensedPath(self.path).data_bind(path=DirectoryDisplay.path)
+        yield DirectoryInput(self.path, select_on_focus=False, compact=True).data_bind(
+            value=DirectoryDisplay.path
+        )
 
 
 class AgentItem(containers.VerticalGroup):
@@ -128,7 +201,7 @@ Your favorite agents.
             return
         child = self.children[self.highlighted]
         assert isinstance(child, LauncherItem)
-        self.post_message(LaunchAgent(child.agent["identity"]))
+        self.post_message(messages.LaunchAgent(child.agent["identity"]))
 
 
 class Launcher(containers.VerticalGroup):
@@ -234,7 +307,7 @@ class AgentGridSelect(GridSelect):
             return
         child = self.children[self.highlighted]
         assert isinstance(child, AgentItem)
-        self.post_message(LaunchAgent(child.agent["identity"]))
+        self.post_message(messages.LaunchAgent(child.agent["identity"]))
 
 
 class Container(containers.VerticalScroll):
@@ -269,11 +342,19 @@ class StoreScreen(Screen):
             key_display="1-9 a-f",
         ),
         Binding("ctrl+r", "resume", "Resume", tooltip="Resume a previous session"),
+        Binding(
+            "ctrl+d",
+            "directory",
+            "Directory",
+            tooltip="Change project directory",
+        ),
     ]
 
     agents_view = getters.query_one("#agents-view", AgentGridSelect)
     launcher = getters.query_one("#launcher", Launcher)
     container = getters.query_one("#container", Container)
+
+    project_dir: reactive[Path] = reactive(Path)
 
     app = getters.app(ToadApp)
 
@@ -286,6 +367,7 @@ class StoreScreen(Screen):
     ):
         self._agents: dict[str, Agent] = {}
         super().__init__(name=name, id=id, classes=classes)
+        self.project_dir = self.app.project_dir
 
     @property
     def agents(self) -> dict[str, Agent]:
@@ -296,7 +378,9 @@ class StoreScreen(Screen):
             with containers.Grid(id="title-grid"):
                 yield Mandelbrot()
                 yield widgets.Label(self.get_info(), id="info")
-
+        yield DirectoryDisplay(self.project_dir).data_bind(
+            project_dir=StoreScreen.project_dir
+        )
         yield Container(id="container", can_focus=False)
         yield widgets.Footer()
 
@@ -392,15 +476,13 @@ class StoreScreen(Screen):
     @on(GridSelect.Selected, ".agents-picker")
     @work
     async def on_grid_select_selected(self, event: GridSelect.Selected):
-        assert isinstance(event.selected_widget, AgentItem)
+        assert isinstance(event.widget, AgentItem)
         from toad.screens.agent_modal import AgentModal
 
-        modal_response = await self.app.push_screen_wait(
-            AgentModal(event.selected_widget.agent)
-        )
+        modal_response = await self.app.push_screen_wait(AgentModal(event.widget.agent))
         self.app.save_settings()
         if modal_response == "launch":
-            self.post_message(LaunchAgent(event.selected_widget.agent["identity"]))
+            self.post_message(messages.LaunchAgent(event.widget.agent["identity"]))
 
     @on(OpenAgentDetails)
     @work
@@ -414,12 +496,12 @@ class StoreScreen(Screen):
         modal_response = await self.app.push_screen_wait(AgentModal(agent))
         self.app.save_settings()
         if modal_response == "launch":
-            self.post_message(LaunchAgent(agent["identity"]))
+            self.post_message(messages.LaunchAgent(agent["identity"]))
 
     @on(GridSelect.Selected, "#launcher GridSelect")
     @work
     async def on_launcher_selected(self, event: GridSelect.Selected):
-        launcher_item = event.selected_widget
+        launcher_item = event.widget
         assert isinstance(launcher_item, LauncherItem)
 
         from toad.screens.agent_modal import AgentModal
@@ -429,44 +511,12 @@ class StoreScreen(Screen):
         )
         self.app.save_settings()
         if modal_response == "launch":
-            self.post_message(LaunchAgent(launcher_item.agent["identity"]))
+            self.post_message(messages.LaunchAgent(launcher_item.agent["identity"]))
 
-    @work
-    async def launch_agent(
-        self,
-        agent_identity: str,
-        agent_session_id: str | None,
-        session_pk: int | None = None,
-    ) -> None:
-        from toad.screens.main import MainScreen
-
-        agent: Agent | None = None
-        if session_pk is not None:
-            db = DB()
-            session = await db.session_get(session_pk)
-            if session is not None:
-                meta = json.loads(session["meta_json"])
-                if agent_data := meta.get("agent_data"):
-                    agent = agent_data
-
-        if agent is None:
-            try:
-                agent = self.agents[agent_identity]
-            except KeyError:
-                self.notify("Agent not found", title="Launch agent", severity="error")
-                return
-        project_path = Path(self.app.project_dir or os.getcwd())
-        screen = MainScreen(
-            project_path, agent, agent_session_id, session_pk=session_pk
-        ).data_bind(
-            column=ToadApp.column,
-            column_width=ToadApp.column_width,
-        )
-        await self.app.push_screen_wait(screen)
-
-    @on(LaunchAgent)
-    def on_launch_agent(self, message: LaunchAgent) -> None:
-        self.launch_agent(message.identity, message.session_id, message.pk)
+    @on(ChangeDirectory)
+    def on_change_directory(self, event: ChangeDirectory) -> None:
+        self.project_dir = Path(event.path)
+        self.app.project_dir = self.project_dir
 
     @work
     async def on_mount(self) -> None:
@@ -528,12 +578,16 @@ class StoreScreen(Screen):
         session = await self.app.push_screen_wait(SessionResumeModal())
         if session is not None:
             self.post_message(
-                LaunchAgent(
+                messages.LaunchAgent(
                     session["agent_identity"],
                     session["agent_session_id"],
                     pk=session["id"],
                 )
             )
+
+    async def action_directory(self) -> None:
+        if (directory_display := self.query_one_optional(DirectoryDisplay)) is not None:
+            directory_display.focus()
 
 
 if __name__ == "__main__":